Fixed listing delete code

+++ database-upgrade/newer-versions/db-40.cfc
@@ -10,6 +10,11 @@
 <cffunction name="executeUpgrade" localmode="modern" access="public" returntype="boolean">
 	<cfargument name="dbUpgradeCom" type="component" required="yes">
 	<cfscript>
+	if(!arguments.dbUpgradeCom.executeQuery(this.datasource, "ALTER TABLE `#request.zos.zcoreDatasourcePrefix#listing_delete`   
+  	ADD  INDEX `NewIndex1` (`listing_id`, `listing_delete_deleted`)")){
+		return false;
+	} 
+
 	if(!arguments.dbUpgradeCom.executeQuery(this.datasource, "ALTER TABLE `#request.zos.zcoreDatasourcePrefix#listing_track`   
 	ADD COLUMN `listing_track_inactive` CHAR(1) DEFAULT '0'   NOT NULL AFTER `listing_track_processed_datetime`,
 	DROP INDEX `NewIndex1`,
